Refatorando a classe Gateway e colocando padrões de projetos melhores de implementação de Interfaces
