+++ app/src/main/java/ifsp/edu/br/minguini/matheus/trabalhoFinal2/fragment/MainFragment.kt
@@ -2,17 +2,15 @@ package ifsp.edu.br.minguini.matheus.trabalhoFinal2.fragment
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
-import android.text.Editable
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import com.squareup.picasso.Picasso
-import ifsp.edu.br.minguini.matheus.trabalhoFinal2.Dto.MovieDTO
 import ifsp.edu.br.minguini.matheus.trabalhoFinal2.Gateway.OmdbGateway
 import ifsp.edu.br.minguini.matheus.trabalhoFinal2.MainActivity
 import ifsp.edu.br.minguini.matheus.trabalhoFinal2.R
+import ifsp.edu.br.minguini.matheus.trabalhoFinal2.dto.MovieDTO
 import ifsp.edu.br.minguini.matheus.trabalhoFinal2.interfaces.MovieCallbackInterface
 import kotlinx.android.synthetic.main.main_fragment.view.*
 
@@ -35,7 +33,9 @@ class MainFragment : Fragment() {
             layoutView.movieIdEditText.text.clear()
         })
 
-        moviesAPI.callback = object : MovieCallbackInterface {
+
+
+        moviesAPI.implementation.callback = object : MovieCallbackInterface {
             override fun onResponse(obj: MovieDTO) {
                 val url = obj.poster
                 if (url != null) {
