+++ app/src/main/java/ifsp/edu/br/minguini/matheus/trabalhoFinal2/interfaces/MovieCallbackInterface.kt
@@ -1,6 +1,6 @@
 package ifsp.edu.br.minguini.matheus.trabalhoFinal2.interfaces
 
-import ifsp.edu.br.minguini.matheus.trabalhoFinal2.Dto.MovieDTO
+import ifsp.edu.br.minguini.matheus.trabalhoFinal2.dto.MovieDTO
 
 interface MovieCallbackInterface {
     fun onResponse(obj: MovieDTO)
